make the guests small integer

+++ database/migrations/2023_04_03_175233_create_waiting_lists_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 ->unsigned()
                 ->nullable();
 
-            $table->tinyInteger('guests')->unsigned();
+            $table->smallInteger('guests')->unsigned();
             $table->dateTime('from_time');
             $table->dateTime('to_time');
